moved function pointer def to mothra_api
